consensus,mobile,params: more per-chain extraction
